Accomodate single or multi-valued relatedResource properties of the collection
Currently all our metadata fields are multivalued. Ryan asked if this method would return an array causing type mismatches (expecting a string), and the answer is yes. I altered the code so that it expects an array of metadata, but will not fail if in the future we make this metadata single valued.